backend/requirements.txt
# ENV
requests
python-dotenv

#google-generativeai

# API
flask
flask-cors
flask-httpauth